AIScore 자료형 변경 및 메인페이지 요약 일정 추가

+++ lib/screens/category_place_page.dart
@@ -87,7 +87,7 @@ class _CategoryPlaceListPageState extends State<CategoryPlaceListPage> {
           final statusDescription = place['status_description'];
           final visitorReviewScore = place['visitorReviewScore'];
           final visitorReviewCount = place['visitorReviewCount'];
-          final int aiScore = place['AI_score'];
+          final double aiScore = place['AI_score'];
 
           double rating = 0;
           if (visitorReviewScore != null) {
